% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_functions.R
\name{getProtID,ProteinCodingGene-method}
\alias{getProtID,ProteinCodingGene-method}
\title{Get Protein ID Method for ProteinCodingGene Class}
\usage{
\S4method{getProtID}{ProteinCodingGene}(geneobject)
}
\arguments{
\item{geneobject}{An object of class ProteinCodingGene}
}
\value{
A character string representing the Protein ID
}
\description{
Retrieves the Protein ID from a ProteinCodingGene object
}
\examples{
library(GenomicRanges)
library(Biostrings)
library(IRanges)

pcg <- ProteinCodingGene(
  geneID = "ENST000001",
  h_symbol = "ACTB",
  geneName = "Beta Actin",
  description = "A major component of the cytoskeleton",
  geneStructure = GenomicRanges::GRanges(
    seqnames = "chr7",
    ranges = IRanges::IRanges(start = 5527175, end = 5535161),
    strand = "+"
  ),
  RnaID = "ENST_RNA_001",
  RnaSEQ = RNAString("AUGGCUAGCUAGCUAGCUA"),
  proteinID = "ENSP000001",
  proteinSEQ = AAString("MEEEIAALVIDNGSG")
)

prot_id <- getProtID(pcg)
print(prot_id)  
}
