% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_functions.R
\name{getRnaID}
\alias{getRnaID}
\title{Get RNA ID Generic Function}
\usage{
getRnaID(geneobject)
}
\arguments{
\item{geneobject}{An object of class Gene}
}
\value{
A character string representing the RNA ID
}
\description{
Retrieves the RNA ID from a Gene object
}
\examples{
library(GenomicRanges)
library(Biostrings)
library(IRanges)

pcg <- ProteinCodingGene(
  geneID = "ENST000001",
  h_symbol = "ACTB",
  geneName = "Beta Actin",
  description = "A major component of the cytoskeleton",
  geneStructure = GenomicRanges::GRanges(
    seqnames = "chr7",
    ranges = IRanges::IRanges(start = 5527175, end = 5535161),
    strand = "+"
  ),
  RnaID = "ENST_RNA_001",
  RnaSEQ = RNAString("AUGGCUAGCUAGCUAGCUA"),
  proteinID = "ENSP000001",
  proteinSEQ = AAString("MEEEIAALVIDNGSG")
)

getRnaID(pcg)
}
