% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_functions.R
\name{getSeedSEQ}
\alias{getSeedSEQ}
\title{Get miRNAs Seed Sequence Generic Function}
\usage{
getSeedSEQ(geneobject)
}
\arguments{
\item{geneobject}{An object of class miRNAGene}
}
\value{
An object of class RNAString representing the miRNA seed sequence
}
\description{
Retrieves the miRNA seed sequence from a miRNAGene object
}
\examples{
library(GenomicRanges)
library(Biostrings)
library(IRanges)

mir <- miRNAGene(
  geneID = "ENST000001",
  h_symbol = "ACTB",
  geneName = "Beta Actin",
  description = "A major component of the cytoskeleton",
  geneStructure = GenomicRanges::GRanges(
    seqnames = "chr7",
    ranges = IRanges::IRanges(start = 5527175, end = 5535161),
    strand = "+"
  ),
  RnaID = "ENST_RNA_001",
  RnaSEQ = RNAString("AUGGCUAGCUAGCUAGCUA"),
  mirna_seed_SEQ= RNAString("UAGCUA")
)

getSeedSEQ(mir)
}
