% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_functions.R
\name{lengthProduct,lncRNAGene-method}
\alias{lengthProduct,lncRNAGene-method}
\title{Length Product Method for lncRNAGene Class}
\usage{
\S4method{lengthProduct}{lncRNAGene}(geneobject)
}
\arguments{
\item{geneobject}{An object of class lncRNAGene}
}
\value{
An integer representing the length of the RNA sequence
}
\description{
Computes the length of the RNA sequence in a lncRNAGene object
}
\examples{
library(GenomicRanges)
library(Biostrings)
library(IRanges)

lnc <- lncRNAGene(
  geneID = "ENST000001",
  h_symbol = "ACTB",
  geneName = "Beta Actin",
  description = "A major component of the cytoskeleton",
  geneStructure = GenomicRanges::GRanges(
    seqnames = "chr7",
    ranges = IRanges::IRanges(start = 5527175, end = 5535161),
    strand = "+"
  ),
  RnaID = "ENST_RNA_001",
  RnaSEQ = RNAString("AUGGCUAGCUAGCUAGCUA")
)

len_lnc <- lengthProduct(lnc)
print(len_lnc)  
}
