% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_functions.R
\name{lengthProduct,miRNAGene-method}
\alias{lengthProduct,miRNAGene-method}
\title{Length Product Method for miRNAGene Class}
\usage{
\S4method{lengthProduct}{miRNAGene}(geneobject)
}
\arguments{
\item{geneobject}{An object of class miRNAGene}
}
\value{
An integer representing the length of the RNA sequence
}
\description{
Computes the length of the RNA sequence in a miRNAGene object
}
\examples{
library(GenomicRanges)
library(Biostrings)
library(IRanges)

mir <- miRNAGene(
  geneID = "ENST000001",
  h_symbol = "ACTB",
  geneName = "Beta Actin",
  description = "A major component of the cytoskeleton",
  geneStructure = GenomicRanges::GRanges(
    seqnames = "chr7",
    ranges = IRanges::IRanges(start = 5527175, end = 5535161),
    strand = "+"
  ),
  RnaID = "ENST_RNA_001",
  RnaSEQ = RNAString("AUGGCUAGCUAGCUAGCUA"),
  mirna_seed_SEQ= RNAString("UAGCUA")
)

len_mir <- lengthProduct(mir)
print(len_mir)  
}
