% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_functions.R
\name{lengthProduct}
\alias{lengthProduct}
\title{Length Product Generic Function}
\usage{
lengthProduct(geneobject)
}
\arguments{
\item{geneobject}{An object of class Gene or from one subclass derived from it}
}
\value{
An integer representing the length of the gene's main product
}
\description{
Computes the length of the gene's main product, that varies based on the gene type:
\itemize{
\item ProteinCodingGene: Number of amino acids in the protein sequence
\item lncRNAGene: Length of the lncRNA sequence
\item miRNAGene: Length of the miRNA sequence
}
}
\examples{
library(GenomicRanges)
library(Biostrings)
library(IRanges)

pcg <- ProteinCodingGene(
  geneID = "ENST000001",
  h_symbol = "ACTB",
  geneName = "Beta Actin",
  description = "A major component of the cytoskeleton",
  geneStructure = GenomicRanges::GRanges(
    seqnames = "chr7",
    ranges = IRanges::IRanges(start = 5527175, end = 5535161),
    strand = "+"
  ),
  RnaID = "ENST_RNA_001",
  RnaSEQ = RNAString("AUGGCUAGCUAGCUAGCUA"),
  proteinID = "ENSP000001",
  proteinSEQ = AAString("MEEEIAALVIDNGSG")
)

lengthProduct(pcg)

lnc <- lncRNAGene(
  geneID = "ENST000001",
  h_symbol = "ACTB",
  geneName = "Beta Actin",
  description = "A major component of the cytoskeleton",
  geneStructure = GenomicRanges::GRanges(
    seqnames = "chr7",
    ranges = IRanges::IRanges(start = 5527175, end = 5535161),
    strand = "+"
  ),
  RnaID = "ENST_RNA_001",
  RnaSEQ = RNAString("AUGGCUAGCUAGCUAGCUA")
)

lengthProduct(lnc)

mir <- miRNAGene(
  geneID = "ENST000001",
  h_symbol = "ACTB",
  geneName = "Beta Actin",
  description = "A major component of the cytoskeleton",
  geneStructure = GenomicRanges::GRanges(
    seqnames = "chr7",
    ranges = IRanges::IRanges(start = 5527175, end = 5535161),
    strand = "+"
  ),
  RnaID = "ENST_RNA_001",
  RnaSEQ = RNAString("AUGGCUAGCUAGCUAGCUA"),
  mirna_seed_SEQ= RNAString("UAGCUA")
)

lengthProduct(mir)
}
