% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_gene_classes.R
\docType{class}
\name{lncRNAGene-class}
\alias{lncRNAGene-class}
\title{lncRNAGene Class (S4 class object)}
\description{
An S4 class representing long non-coding RNA (lncRNA) genes, inheriting from the virtual Gene class.
It does not include additional slots beyond those inherited from Gene since the RNA is it's final product.
}
\examples{
library(GenomicRanges)
library(Biostrings)
library(IRanges)

lncr_object <- new("lncRNAGene",
           geneID = "ENST000002",
           h_symbol = "LNC1",
           geneName = "Long Non-Coding RNA 1",
           description = "A long non-coding RNA gene involved in regulation",
           geneStructure = GenomicRanges::GRanges(seqnames = "chr2",
                                    ranges = IRanges::IRanges(start = 1500000, end = 1501000),
                                    strand = "-"),
           RnaID = "ENST_RNA_002",
           RnaSEQ = RNAString("AUCGAUCGAUGCA"))

}
\seealso{
\code{\linkS4class{Gene}}
}
\author{
Alexander Dell'Orti \cr Politecnico di Milano\cr Maintainer: Alexander
Dell'Orti \cr E-Mail: \href{mailto:alexander.dellorti@mail.polimi.it}{alexander.dellorti@mail.polimi.it}
}
