% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_constructors.R
\name{lncRNAGene}
\alias{lncRNAGene}
\title{Create a lncRNAGene Object}
\usage{
lncRNAGene(
  geneID,
  h_symbol,
  geneName,
  description,
  geneStructure,
  RnaID,
  RnaSEQ
)
}
\arguments{
\item{geneID}{Character, Ensembl or NCBI Gene identifier}

\item{h_symbol}{Character, Gene HUGO symbol}

\item{geneName}{Character, Full gene name}

\item{description}{Character, Gene description}

\item{geneStructure}{GRanges, Genomic Ranges object representing the gene's structure}

\item{RnaID}{Character, RNA identifier}

\item{RnaSEQ}{RNAString, RNA sequence}
}
\value{
lncRNAGene S4 object
}
\description{
This constructor function creates a new lncRNAGene class object with input validation.
If any of the attributes are missing, the default value defined by the class prototype parameter
will be assigned and a warning message will be printed.
}
\examples{
# Load necessary libraries
library(GenomicRanges)
library(Biostrings)
library(IRanges)

# Creating a lncRNAGene object (with all attributes)
lnc_object <- lncRNAGene(
  geneID = "ENST000001",
  h_symbol = "LNC1",
  geneName = "Long Non-Coding RNA 1",
  description = "A long non-coding RNA gene involved in regulation",
  geneStructure = GenomicRanges::GRanges(
    seqnames = "chr2",
    ranges = IRanges::IRanges(start = 1500000, end = 1501000),
    strand = "-"
  ),
  RnaID = "ENST_RNA_002",
  RnaSEQ = Biostrings::RNAString("AUCGAUCGAUGCA")
)

}
\seealso{
\code{\linkS4class{Gene}}, \code{\linkS4class{ProteinCodingGene}}, \code{\linkS4class{miRNAGene}}
}
\author{
Alexander Dell'Orti \cr Politecnico di Milano\cr Maintainer: Alexander
Dell'Orti \cr E-Mail: \href{mailto:alexander.dellorti@mail.polimi.it}{alexander.dellorti@mail.polimi.it}
}
