% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_gene_classes.R
\docType{class}
\name{miRNAGene-class}
\alias{miRNAGene-class}
\title{miRNAGene Class (S4 class object)}
\description{
An S4 class representing microRNA (miRNA) genes, inheriting from the virtual Gene class.
It includes a slot specific to miRNA seed sequences.
}
\section{Slots}{

\describe{
\item{\code{mirna_seed_SEQ}}{RNAString, miRNA seed sequence in RNAString format}
}}

\examples{
library(GenomicRanges)
library(Biostrings)
library(IRanges)

mir_object <- new("miRNAGene",
           geneID = "ENST000003",
           h_symbol = "MIR1",
           geneName = "MicroRNA 1",
           description = "A microRNA gene involved in post-transcriptional regulation",
           geneStructure = GenomicRanges::GRanges(seqnames = "chr3",
                                    ranges = IRanges::IRanges(start = 2500000, end = 2500500),
                                    strand = "+"),
           RnaID = "ENST_RNA_003",
           RnaSEQ = RNAString("AUGCUAGCUA"),
           mirna_seed_SEQ = RNAString("AUCGUAA"))

}
\seealso{
\code{\linkS4class{Gene}}
}
\author{
Alexander Dell'Orti \cr Politecnico di Milano\cr Maintainer: Alexander
Dell'Orti \cr E-Mail: \href{mailto:alexander.dellorti@mail.polimi.it}{alexander.dellorti@mail.polimi.it}
}
