% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_gene_classes.R
\docType{class}
\name{ProteinCodingGene-class}
\alias{ProteinCodingGene-class}
\title{ProteinCodingGene Class (S4 class object)}
\description{
An S4 class representing protein-coding genes, inheriting from the virtual Gene class.
It includes additional slots specific to protein products.
}
\section{Slots}{

\describe{
\item{\code{proteinID}}{Character, Protein ID}

\item{\code{proteinSEQ}}{AAString, Protein amino acid sequence in AAString format}
}}

\examples{
library(GenomicRanges)
library(Biostrings)
library(IRanges)

pcg_object <- new("ProteinCodingGene",
           geneID = "ENST000001",
           h_symbol = "ACTB",
           geneName = "Beta Actin",
           description = "A major component of the cytoskeleton",
           geneStructure = GenomicRanges::GRanges(seqnames = "chr7",
                                    ranges = IRanges::IRanges(start = 5527175, end = 5535161),
                                    strand = "+"),
           RnaID = "ENST_RNA_001",
           RnaSEQ = RNAString("AUGGCUAGCUAGCUAGCUA"),
           proteinID = "ENSP000001",
           proteinSEQ = AAString("MEEEIAALVIDNGSG"))

}
\seealso{
\code{\linkS4class{Gene}}
}
\author{
Alexander Dell'Orti \cr Politecnico di Milano\cr Maintainer: Alexander
Dell'Orti \cr E-Mail: \href{mailto:alexander.dellorti@mail.polimi.it}{alexander.dellorti@mail.polimi.it}
}
