% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_functions.R
\name{setGeneID}
\alias{setGeneID}
\title{Set Gene ID Generic Function}
\usage{
setGeneID(geneobject, newid)
}
\arguments{
\item{geneobject}{An object of class Gene}

\item{newid}{A character string representing the new Gene ID}
}
\value{
The updated Gene object with the new Gene ID
}
\description{
Sets a new Gene ID for a Gene object
}
\examples{
library(GenomicRanges)
library(Biostrings)
library(IRanges)

pcg <- ProteinCodingGene(
  geneID = "ENST000001",
  h_symbol = "ACTB",
  geneName = "Beta Actin",
  description = "A major component of the cytoskeleton",
  geneStructure = GenomicRanges::GRanges(
    seqnames = "chr7",
    ranges = IRanges::IRanges(start = 5527175, end = 5535161),
    strand = "+"
  ),
  RnaID = "ENST_RNA_001",
  RnaSEQ = RNAString("AUGGCUAGCUAGCUAGCUA"),
  proteinID = "ENSP000001",
  proteinSEQ = AAString("MEEEIAALVIDNGSG")
)

pcg <- setGeneID(pcg, "ENST000002")
gene_id <- getGeneID(pcg)
print(gene_id)  
}
